Improve cache mechanism by not throwing away the repo

+++ eng/download-mlaunch.ps1
@@ -1,3 +1,9 @@
+<#
+    Commandlet downloads specific revision of the mlaunch binary from xamarin/macios-binaries repository.
+    This binary is then bundled with XHarness inside the NuGet.
+
+    Revision is cached in a temp dir and re-used for new builds.
+#>
 [CmdletBinding(PositionalBinding=$false)]
 Param(
     [ValidateNotNullOrEmpty()]
@@ -13,7 +19,7 @@ Param(
 
 New-Item -Path $TargetDir -ItemType Directory -ErrorAction SilentlyContinue
 
-Write-Host "mlaunch revision $commit will be installed into $TargetDir"
+Write-Host "Getting mlaunch revision $commit into $TargetDir"
 
 $tagFile = Join-Path $TargetDir ".tag-$commit"
 
@@ -23,8 +29,28 @@ if (Test-Path $tagFile) {
 }
 
 $binariesRepo = Join-Path $TempDir "macios-binaries"
+$tagFileInRepo = Join-Path $binariesRepo ".tag-$commit"
 
-Remove-Item -Recurse -Force -ErrorAction SilentlyContinue $binariesRepo
+if (Test-Path $binariesRepo) {
+    if (Test-Path $tagFileInRepo) {
+        $path = Join-Path $binariesRepo "mlaunch"
+
+        # Copy mlaunch to the target folder
+        Copy-Item -Path $path -Destination $TargetDir -Recurse -Verbose
+
+        # Tag the version of mlaunch we have
+        New-Item -ItemType File -Path $tagFile
+
+        Write-Host "Finished downloading mlaunch from cache"
+        ExitWithExitCode 0
+    }
+
+    Write-Host "Found cached repository but different version was checked out. Downloading again for $commit..."
+    Remove-Item -Recurse -Force -ErrorAction SilentlyContinue $binariesRepo
+}
+
+Write-Host "Cloning the xamarin-binaries repository. This will take few minutes.."
+
 New-Item -ItemType Directory -Force -ErrorAction Stop -Path $binariesRepo
 
 # Shallow-clone the xamarin/macios-binaries repo
@@ -37,8 +63,10 @@ Set-Content -Path ".git/info/sparse-checkout" -Value "mlaunch"
 git fetch --depth 1 origin $commit
 git checkout FETCH_HEAD
 
-# Copy mlaunch to the artifacts folder
-Move-Item -Path "mlaunch" -Destination $TargetDir -Verbose
+New-Item -ItemType File -Path $tagFileInRepo
+
+# Copy mlaunch to the target folder
+Copy-Item -Path "mlaunch" -Destination $TargetDir -Recurse -Verbose
 
 # Tag the version of mlaunch we have
 New-Item -ItemType File -Path $tagFile
